freemocap_blender_addon/requirements.txt
ruamel.yaml
numpy